WIP: remove cruft
