+++ lib/magic-link.js
@@ -1,32 +1,8 @@
 "use strict";
 
-let crypto = require("crypto");
-
 let E = require("./errors.js");
 
-let localhosts = ["::ffff:127.0.0.1", "127.0.0.1", "::1"];
-
 let Challenger = module.exports;
-
-/**
- * @param {import('express').Handler} fn
- * @param {import('express').Request} req
- * @param {import('express').Response} res
- * @returns {Promise<any>}
- */
-async function promisifyHandler(fn, req, res) {
-  let err = await new Promise(async function (resolve, reject) {
-    try {
-      //@ts-ignore
-      await fn(req, res, resolve).catch(reject);
-    } catch (err) {
-      reject(err);
-    }
-  });
-  if (err) {
-    throw err;
-  }
-}
 
 /**
  * @param {any} opts
@@ -44,105 +20,9 @@ Challenger._createRoutes = function ({
   iss = "",
   //@ts-ignore
   verifier,
-  //@ts-ignore
-  notify,
-  DEVELOPMENT = false,
-  _developmentSendSecretVerificationCode = false,
-
   authnParam = "authn",
-  /** @type {function} */
-  //@ts-ignore
-  _strategyHandler,
-  /** @type {function} */
-  //@ts-ignore
-  grantTokensAndCookie,
 }) {
   let Routes = {};
-
-  //
-  // Email Verification Challenges
-  //
-  /** @type {import('express').Handler} */
-  Routes.orderVerification = async function (req, res) {
-    // arbitrary client state (so that when verification request is sent from
-    // desktop and responded to on mobile or vice-versa the client can rehydrate
-    // something meaningful)
-    let state = req.body?.state;
-    if (state && "object" === typeof state) {
-      let maxKeys = 20;
-      let maxSize = 1024;
-      let len = Object.keys(state).length;
-      if (len > maxKeys) {
-        let err = new Error(
-          `'state' has '${len}' keys, but should be limited to '${maxKeys}'`,
-        );
-        //@ts-ignore
-        err.code = "BAD_REQUEST";
-        //@ts-ignore
-        err.status = 400;
-        throw err;
-      }
-      let size = Object.keys(state).reduce(function (n, k) {
-        let val = state[k];
-        if ("string" !== typeof val) {
-          let json = JSON.stringify(val);
-          let err = new Error(
-            `'state' should only have string values, not '${json}'`,
-          );
-          //@ts-ignore
-          err.code = "BAD_REQUEST";
-          //@ts-ignore
-          err.status = 400;
-          throw err;
-        }
-        return n + k.length + val.length;
-      }, 0);
-      if (size > maxSize) {
-        let err = new Error(
-          `'state' should be constrained to ${maxSize} bytes`,
-        );
-        //@ts-ignore
-        err.code = "BAD_REQUEST";
-        //@ts-ignore
-        err.status = 400;
-        throw err;
-      }
-    }
-
-    await promisifyHandler(Routes.order, req, res);
-    // this req must have req[authnParam] as the notifier info
-    await notify(req, {});
-    //await verifier.notify(c, req, opts);
-    //@ts-ignore
-    req[authnParam] = null;
-
-    let result = Object.assign({ success: true }, req.authn.order);
-
-    if (DEVELOPMENT) {
-      // TODO return Promise.resolve().then(doThisRoute).catch(next);
-      let _devSendCode =
-        req.body._developmentSendSecretVerificationCode ||
-        req.body._development_send_secret_verification_code;
-      let devSendCode = _devSendCode && _developmentSendSecretVerificationCode;
-      // the raw connection is local AND the proxied connection is local
-      /*
-      let isLocal =
-        localhosts.includes(res.socket?.remoteAddress || "") &&
-        (!req.ip || localhosts.includes(req.ip) || true);
-      if (devSendCode && isLocal) {
-      */
-      console.warn(
-        "[auth3000] SECURITY: giving out the secret verification code to localhost in DEVELOPMENT mode",
-      );
-
-      result._development_secret_verification_code = req.authn.order.getCode();
-      /*
-      }
-      */
-    }
-
-    res.json(result);
-  };
 
   /** @type {import('express').Handler} */
   Routes.order = async function (req, res, next) {
@@ -261,6 +141,7 @@ Challenger._createRoutes = function ({
       expires_at: c.expires_at,
     };
 
+    //@ts-ignore
     req[authnParam] = {
       id: id,
       code: code,
@@ -272,35 +153,6 @@ Challenger._createRoutes = function ({
       next();
     }
     //return authnParams;
-  };
-
-  /** @type {import('express').Handler} */
-  Routes.redeemCode = async function (req, res) {
-    await promisifyHandler(Routes.useCode, req, res);
-
-    let allClaims = await _strategyHandler(req, res);
-    //@ts-ignore
-    let authn = req[authnParam];
-
-    //@ts-ignore
-    req[authnParam] = null;
-
-    // TODO deprecate, remove
-    if (allClaims || !res.headersSent) {
-      let { id_token, access_token } = await grantTokensAndCookie(
-        allClaims,
-        req,
-        res,
-      );
-
-      res.json({
-        success: true,
-        status: "valid",
-        state: authn.state,
-        id_token: id_token,
-        access_token: access_token,
-      });
-    }
   };
 
   /** @type {import('express').Handler} */
@@ -363,34 +215,6 @@ Challenger._createRoutes = function ({
       canceled_at: c.canceled_at,
       canceled_by: c.canceled_by,
     });
-  };
-
-  /** @type {import('express').Handler} */
-  Routes.exchangeReceipt = async function (req, res) {
-    await promisifyHandler(Routes.useReceipt, req, res);
-
-    let allClaims = await _strategyHandler(req, res);
-    //@ts-ignore
-    let authn = req[authnParam];
-    //@ts-ignore
-    req[authnParam] = null;
-
-    // TODO deprecate
-    if (allClaims || !res.headersSent) {
-      let { id_token, access_token } = await grantTokensAndCookie(
-        allClaims,
-        req,
-        res,
-      );
-
-      res.json({
-        success: true,
-        status: "valid",
-        state: authn.state,
-        id_token: id_token,
-        access_token: access_token,
-      });
-    }
   };
 
   /** @type {import('express').Handler} */
