+++ lib/session.js
@@ -125,12 +125,16 @@ let defaultTrustedMaxAge = "7d";
  * @param {JwsPriv} PRIVATE_KEY
  * @param {Object} myOptions
  * @param {String} myOptions.cookiePath
- * @param {Boolean} myOptions.DEVELOPMENT
  */
 LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
   let opts = {
     issuer: issuer,
     oidc: {},
+    credentials: {
+      basic: true,
+      username: "username",
+      password: "password",
+    },
     challenge: {
       idByteCount: 4,
       /** @type {import('crypto').BinaryToTextEncoding} */
@@ -154,7 +158,6 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
     challengeMaxAttempts: 0,
     authnParam: "authn",
     cookiePath: myOptions.cookiePath,
-    DEVELOPMENT: myOptions.DEVELOPMENT,
     // assigned elsewhere
     secret: "",
   };
@@ -185,15 +188,6 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
     throw new Error(
       `[libauth] 'opts.cookiePath' should be scoped to a path, such as '/api/authn/'`,
     );
-  }
-
-  // Dev / Localhost Stuff
-  if (opts.DEVELOPMENT) {
-    console.info(
-      "[libauth] [ENV=DEVELOPMENT] Allow Insecure (localhost) Cookies",
-    );
-    // allow non-https cookies
-    cookieDefaults.secure = false;
   }
 
   /*
@@ -286,12 +280,6 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
   let a3k = {
     /** @param {any} _opts // TODO */
     challenge: function (_opts) {
-      // it's a verifier, not an options object
-      if (_opts && _opts.setDefaults) {
-        opts.verifier = _opts;
-        return;
-      }
-
       opts.challenge = Object.assign(opts.challenge, _opts);
 
       if (!opts.challenge.store && !opts.store) {
@@ -343,20 +331,20 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
 
     /** @param {any} _opts // TODO */
     credentials: function (_opts) {
-      opts = Object.assign(opts, _opts);
+      let credOpts = Object.assign(opts.credentials, _opts);
 
       /** @type {import('express').Handler} */
       async function _credentialRoutes(req, res, next) {
         let creds;
 
         let authBasic = req.headers["authorization"] || "";
-        if (false !== _opts?.basic && authBasic.startsWith("Basic ")) {
+        if (false !== credOpts.basic && authBasic.startsWith("Basic ")) {
           creds = Util.decodeAuthorizationBasic(req);
         }
 
         if (!creds?.username) {
-          let userKey = _opts?.username || _opts?.user || "username";
-          let passKey = _opts?.password || _opts?.pass || "password";
+          let userKey = credOpts.username || "username";
+          let passKey = credOpts.password || "password";
           creds = {
             username: req.body[userKey],
             password: req.body[passKey],
