+++ lib/util.js
@@ -73,3 +73,29 @@ function secureCompare(trusted = "", given = "", min = 16) {
 
   return crypto.timingSafeEqual(Buffer.from(trusted), Buffer.from(given));
 }
+
+/**
+ * @param {import('express').Handler} fn
+ * @param {import('express').Request} req
+ * @param {import('express').Response} res
+ * @returns {Promise<any>}
+ */
+/*
+async function promisifyHandler(fn, req, res) {
+  let err = await new Promise(async function (resolve, reject) {
+    try {
+      //@ts-ignore
+      await fn(req, res, resolve).catch(reject);
+    } catch (err) {
+      reject(err);
+    }
+  });
+  if (err) {
+    throw err;
+  }
+}
+*/
+
+/*
+let localhosts = ["::ffff:127.0.0.1", "127.0.0.1", "::1"];
+*/
