+++ lib/verifier.js
@@ -317,6 +317,8 @@ C.create = function ({
    * @param {String} opts.secret
    * @param {String} opts.authnParam
    */
+  // for internal use by libauth
+  // (so that issuer doesn't have to be set manually)
   Challenge.setDefaults = function (opts) {
     if (!iss && opts.iss) {
       iss = opts.iss;
